Running notes overhaul for new run notes database: more work need for workset notes
