+++ run_dir/static/js/flowcell.js
@@ -19,7 +19,6 @@ $('.lane-copy').click(function(){
 
 // from runnin_notes.js
 // these vars are needed for load_running_notes(); -> it's js
-var notetype='flowcell';
 var flowcell_id_reference=$('#flowcells-js').attr('data-flowcell'); // passing parameters to <script> from flowcell.html
 load_running_notes();
 // from links.js
